move from platform/kubernetes/nodeinfo to metric/hostinfo, to share this implementation with ECS
